Make yarn and package log files text for git
yarn.lock (and package-lock.json which is not used in this project) is a text file.
So it could be useful to consider it like that in order to enjoy diff views.
Git usually guesses correctly file types, so only exceptions requires to defined such attributes.

Note: 'text -diff' consider files as text, but prevent any diff view.