chore: schedule dead link checker every 6 months
Dead link checker raise a lot of false positives (links which are not available temporarily).
More over, SEO impacts is not proven (does the source site is impected or the destination site?).

So, for all these reasons, we decided to decrease the check frequency to every 6 months.

Why not `0 10 1 */6 *`? To check it is triggered the day after this commit.